Sync

+++ dango/httpd/src/graphql/types/transfer.rs
@@ -10,7 +10,7 @@ use {
 #[graphql(complex)]
 #[serde(default)]
 pub struct Transfer {
-    pub block_height: i64,
+    pub block_height: u64,
     pub created_at: DateTime<Utc>,
     pub from_address: String,
     pub to_address: String,
@@ -21,7 +21,7 @@ pub struct Transfer {
 impl From<entity::transfers::Model> for Transfer {
     fn from(item: entity::transfers::Model) -> Self {
         Self {
-            block_height: item.block_height,
+            block_height: item.block_height as u64,
             created_at: Utc.from_utc_datetime(&item.created_at),
             from_address: item.from_address,
             to_address: item.to_address,
